requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
Janome==0.5.0
multidict==6.0.4
mysqlclient==2.2.1
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
yarl==1.9.4
